chore: code clean-up
