+++ package.json
@@ -19,7 +19,6 @@
     "@nubjs/nub": "0.7.5",
     "@types/node": "26.2.0",
     "eslint-plugin-legibility": "^0.3.0",
-    "fast-check": "^4.9.0",
     "oxfmt": "^0.63.0",
     "oxlint": "^1.78.0",
     "pastoralist": "1.13.0",
@@ -27,7 +26,6 @@
     "release-it": "21.0.2",
     "rolldown": "1.2.3",
     "scriptc": "0.0.26",
-    "tsx": "4.23.1",
     "typescript": "^7.0.2"
   },
   "exports": {
@@ -75,7 +73,7 @@
     "clean": "nub run clean-dist && cd page/app && rm -rf dist",
     "cmd:node:test-fail-run": "node dist/cli.js --config './tests/unit/fixtures/test-fail-package.json' --files 'test-fail-package.json' --rootDir './tests/unit/fixtures/' --debug",
     "cmd:node:test-pass-run": "node dist/cli.js --config './tests/unit/fixtures/test-pass-package.json' --files 'test-pass-package.json'  --rootDir './tests/unit/fixtures/' --debug",
-    "coverage": "tsx --experimental-test-module-mocks --experimental-test-coverage --test --test-coverage-include='src/**/*.ts' --test-coverage-lines=80 --test-reporter=spec --test-reporter-destination=stdout --test-reporter=lcov --test-reporter-destination=coverage/lcov.info",
+    "coverage": "node scripts/run/index.ts --coverage",
     "dev": "nub run docs:dev",
     "docs:dev": "cd page/app && nub run dev",
     "docs:build": "cd page/app && nub run build",
@@ -95,24 +93,24 @@
     "prepare": "nub run hooks:setup",
     "prepublishOnly": "nub run clean-dist && nub run build-dist",
     "preview": "nub run docs:preview",
-    "release": "nub scripts/release.ts",
-    "release:patch": "nub scripts/release.ts --increment=patch",
-    "release:patch:dry": "nub scripts/release.ts --increment=patch --dry-run",
-    "release:minor": "nub scripts/release.ts --increment=minor",
-    "release:minor:dry": "nub scripts/release.ts --increment=minor --dry-run",
-    "release:major": "nub scripts/release.ts --increment=major",
-    "release:major:dry": "nub scripts/release.ts --increment=major --dry-run",
-    "release:alpha": "nub scripts/release.ts --preRelease=alpha",
-    "release:alpha:dry": "nub scripts/release.ts --preRelease=alpha --dry-run",
-    "release:beta": "nub scripts/release.ts --preRelease=beta",
-    "release:beta:dry": "nub scripts/release.ts --preRelease=beta --dry-run",
-    "release:rc": "nub scripts/release.ts --preRelease=rc",
-    "release:rc:dry": "nub scripts/release.ts --preRelease=rc --dry-run",
-    "release:tag": "nub scripts/tag-release.ts",
-    "release:tag:dry": "nub scripts/tag-release.ts --dry-run",
-    "release:dry": "nub scripts/release.ts --dry-run",
+    "release": "nub scripts/release/index.ts",
+    "release:patch": "nub scripts/release/index.ts --increment=patch",
+    "release:patch:dry": "nub scripts/release/index.ts --increment=patch --dry-run",
+    "release:minor": "nub scripts/release/index.ts --increment=minor",
+    "release:minor:dry": "nub scripts/release/index.ts --increment=minor --dry-run",
+    "release:major": "nub scripts/release/index.ts --increment=major",
+    "release:major:dry": "nub scripts/release/index.ts --increment=major --dry-run",
+    "release:alpha": "nub scripts/release/index.ts --preRelease=alpha",
+    "release:alpha:dry": "nub scripts/release/index.ts --preRelease=alpha --dry-run",
+    "release:beta": "nub scripts/release/index.ts --preRelease=beta",
+    "release:beta:dry": "nub scripts/release/index.ts --preRelease=beta --dry-run",
+    "release:rc": "nub scripts/release/index.ts --preRelease=rc",
+    "release:rc:dry": "nub scripts/release/index.ts --preRelease=rc --dry-run",
+    "release:tag": "nub scripts/release/index.ts tag",
+    "release:tag:dry": "nub scripts/release/index.ts tag --dry-run",
+    "release:dry": "nub scripts/release/index.ts --dry-run",
     "setup": "nub install && nub run hooks:setup && nub run build",
-    "test": "tsx --experimental-test-module-mocks --test",
+    "test": "node scripts/run/index.ts",
     "test:e2e:binary": "./tests/e2e/test-binary.sh",
     "typecheck": "nub run typecheck-src",
     "typecheck-src": "tsc --noEmit",
