+++ scripts/release/constants.ts
@@ -1,0 +1,13 @@
+import type { PreRelease, ReleaseIncrement } from "./types";
+
+export const TAG_VERSION_PATTERN = /^\d+\.\d+\.\d+(-[0-9A-Za-z.-]+)?(\+[0-9A-Za-z.-]+)?$/;
+export const RELEASE_VERSION_PATTERN = /\d+\.\d+\.\d+(?:-[0-9A-Za-z.-]+)?(?:\+[0-9A-Za-z.-]+)?/g;
+export const STABLE_VERSION_PATTERN = /^\d+\.\d+\.\d+$/;
+export const PRE_RELEASE_VERSION_PATTERN = /^\d+\.\d+\.\d+-[0-9A-Za-z.-]+(?:\+[0-9A-Za-z.-]+)?$/;
+export const PRE_RELEASES = new Set<PreRelease>(["alpha", "beta", "rc"]);
+export const RELEASE_INCREMENTS = new Set<ReleaseIncrement>(["patch", "minor", "major"]);
+export const SAFE_SHELL_ARG_PATTERN = /^[A-Za-z0-9_./:=@-]+$/;
+export const DEFAULT_RELEASE_TIMEOUT_MINUTES = 90;
+export const RELEASE_POLL_INTERVAL_MS = 30_000;
+export const COMMIT_PATTERN = /^[0-9a-f]{40}$/i;
+export const REMOVED_VERSION_LINE_PATTERN = /^-\s*"version":\s*"[^"]+",\s*$/;
