+++ scripts/run/types.ts
@@ -1,0 +1,4 @@
+export interface TestRunnerOptions {
+  coverageEnabled: boolean;
+  testArgs: string[];
+}
